# rudi-microservice/rudi-microservice-acl/openapi/rudi-acl-api.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "API Public ACL Rudi",
		"version": "0.0.1",
		"title": "rudi-acl-api",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "contact@rudi.fr"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"servers": [
		{
			"url": "/{basePath}",
			"description": "Rudi Acl API",
			"variables": {
				"basePath": {
					"default": "acl/v1"
				}
			}
		}
	],
...
	"paths": {
...
		"/account/authenticate": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Demande d'authentification de l'utilisateur",
				"description": "Demande d'authentification de l'utilisateur",
				"operationId": "authenticate1",
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string",
							"description": "L'access token reçu lors de la redirection"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/Tokens"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Demande d'authentification de l'utilisateur",
				"description": "Demande d'authentification de l'utilisateur",
				"operationId": "authenticate2",
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string",
							"description": "L'access token reçu lors de la redirection"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/Tokens"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account/{login}/is-created-not-validated": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Indique si un compte en attente d'activation existe pour le login",
				"description": "Indique si un compte en attente d'activation existe pour le login",
				"operationId": "isCreatedNotValidated",
				"parameters": [
					{
						"in": "path",
						"name": "login",
						"description": "Le login du compte",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account/{login}/must-validate-captcha": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Indique si un compte doit valider le captcha pour se connecter",
				"description": "Indique si un compte doit valider le captcha pour se connecter",
				"operationId": "mustValidateCaptcha",
				"parameters": [
					{
						"in": "path",
						"name": "login",
						"description": "Le login du compte",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account": {
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Création d'un compte utilisateur",
				"description": "Permet à un utilisateur anonyme du site de créer un compte utilisateur",
				"operationId": "requestAccountCreation",
				"requestBody": {
					"description": "Informations pour la création du compte",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/Account"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Validation d'un compte utilisateur",
				"description": "Valide une demande création de compte",
				"operationId": "validateAccount",
				"requestBody": {
					"description": "Informations pour la création du compte",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/User"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account/change-password": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Demande de réinitialisation de mot-de-passe",
				"description": "Génère un token pour réinitialiser le mot-de-passe d'un compte et l'envoie par mail à l'utilisateur.",
				"operationId": "requestPasswordChange",
				"parameters": [
					{
						"in": "query",
						"name": "email",
						"description": "Email de l'utilisateur souhaitant changer de mot-de-passe",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Un token pour changer le mot-de-passe. Pour des raisons de sécurité, on envoie un token factice lorsque l'email n'est pas connu."
					},
					"500": {
						"description": "Erreur technique liée au serveur.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Modification du mot-de-passe d'un compte",
				"description": "Modifie le mot-de-passe d'un compte à partir d'un token. Ce token est obtenu suite à une demande de changement de mot-de-passe. Envoie un mail de confirmation à l'utilisateur concerné",
				"operationId": "validatePasswordChange",
				"requestBody": {
					"description": "Informations pour la modification du mot-de-passe",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/PasswordChange"
							}
						}
					}
				},
				"responses": {
					"204": {
						"description": "OK. Pour des raisons de sécurité, on répond toujours OK, même lorsque le token n'est pas connu."
					},
					"500": {
						"description": "Erreur technique liée au serveur.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account/change-password/check-token": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Vérification du token",
				"description": "Vérifie le token utilisé pour modifier le mot-de-passe d'un compte. Ce token est obtenu suite à une demande de changement de mot-de-passe.",
				"operationId": "checkPasswordChangeToken",
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Le token est valide."
					},
					"500": {
						"description": "Erreur technique liée au serveur.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/account/logout": {
			"delete": {
				"tags": [
					"acl"
				],
				"summary": "Déconexion d'un utilisateur",
				"description": "Permet à un utilisateur de se déconnecter",
				"operationId": "accountLogout",
				"parameters": [
					{
						"in": "header",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string",
							"description": "Token a surpprimer pour confirmer la déconnexion"
						}
					}
				],
				"responses": {
					"204": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/addressRoles": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Collection de roles d'adresses'",
				"description": "Retourne une liste de roles d'adresses",
				"operationId": "searchAddressRoles",
				"parameters": [
					{
						"in": "query",
						"name": "active",
						"description": "Elements actif ou tous",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "type",
						"description": "Type du rôle",
						"required": false,
						"schema": {
							"$ref": "rudi-acl-model.json#/components/schemas/AddressType"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"acl"
				],
				"summary": "Ajout d'un nouveau role d'adresse",
				"description": "Permet d'ajouter un nouveau role d'adresse",
				"operationId": "createAddressRole",
				"requestBody": {
					"description": "info role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"acl"
				],
				"summary": "Mise à jour d'un role d'adresse",
				"description": "Mise à jour d'un role d'adresse",
				"operationId": "updateAddressRole",
				"requestBody": {
					"description": "info role",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/addressRoles/{uuid}": {
			"get": {
				"tags": [
					"acl"
				],
				"summary": "Retourne un role d'adresse",
				"description": "Retourne un role d'adresse",
				"operationId": "getAddressRole",
				"parameters": [
					{
						"name": "uuid",
						"in": "path",
						"description": "uuid role",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "rudi-acl-model.json#/components/schemas/AddressRole"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": [
					"acl"
				],
				"summary": "Suppression d'un role d'adresse",
				"description": "Suppression d'un role d'adresse",
				"operationId": "deleteAddressRole",
				"parameters": [
					{
						"name": "uuid",
						"in": "path",
						"description": "uuid role",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"ApiError": {
				"type": "object",
				"properties": {
					"code": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				}
			},
...
		}
	}
}